DVX-944 Sync with upstream and re-apply our fixes (#11)
Synced with upstream's main branch with our changes re-applied.

---------
