+++ .envrc
@@ -1,0 +1,2 @@
+source_up
+use node v20
