+++ .gitignore
@@ -94,6 +94,6 @@ typings/
 
 # Text editor files
 .vscode/
+.idea/
 src/test.ts
-.idea
 *.iml
