+++ action.yml
@@ -67,8 +67,10 @@ inputs:
 outputs:
   cache-hit:
     description: "A boolean value to indicate an exact match was found for the primary key"
+  cache-size:
+    description: "A integer value denoting the size of the cache object found (measured in bytes)"
 runs:
-  using: node16
+  using: node20
   main: "dist/restore/index.js"
   post: "dist/save/index.js"
   post-if: "success()"
