+++ jest.config.js
@@ -14,7 +14,7 @@ const processStdoutWrite = process.stdout.write.bind(process.stdout);
 process.stdout.write = (str, encoding, cb) => {
   // Core library will directly call process.stdout.write for commands
   // We don't want :: commands to be executed by the runner during tests
-  if (!str.match(/^::/)) {
+  if (!`${str}`.match(/^::/)) {
     return processStdoutWrite(str, encoding, cb);
   }
 };
