+++ package.json
@@ -2,22 +2,23 @@
   "name": "actions-cache",
   "main": "lib/restoreCache.js",
   "license": "MIT",
+  "packageManager": "yarn@1.22.22",
   "scripts": {
-    "build": "tsc && ncc build -o dist/restore src/restore.ts && ncc build -o dist/save src/save.ts",
+    "build": "tsc && ncc build -o dist/restore src/restore.ts && ncc build -o dist/save src/save.ts && ncc build -o dist/saveOnly src/saveOnly.ts",
     "test": "tsc --noEmit && jest --coverage"
   },
   "dependencies": {
-    "@actions/cache": "3.0.6",
-    "@actions/core": "1.10.0",
-    "minio": "7.0.32"
+    "@actions/cache": "^3.2.2",
+    "@actions/core": "^1.10.0",
+    "minio": "^7.1.1"
   },
   "devDependencies": {
-    "@types/jest": "29.2.3",
-    "@types/minio": "7.0.15",
-    "@types/node": "^18.11.9",
-    "@zeit/ncc": "0.22.3",
-    "jest": "29.3.1",
-    "ts-jest": "29.0.3",
-    "typescript": "4.9.3"
+    "@types/jest": "^26.0.23",
+    "@types/minio": "^7.0.7",
+    "@types/node": "^20.5.1",
+    "@vercel/ncc": "^0.36.1",
+    "jest": "^29.6.2",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.1.6"
   }
 }
