+++ restore/action.yml
@@ -1,0 +1,61 @@
+name: "s3 cache"
+author: tespkg
+description: "S3 actions cache with actions/cache@v2 passthrough"
+inputs:
+  endpoint:
+    description: "s3 / minio endpoint, see: https://docs.min.io/docs/javascript-client-api-reference.html"
+    required: false
+    default: s3.amazonaws.com
+  region:
+    description: "AWS region"
+    required: false
+  port:
+    description: minio port
+    required: false
+  insecure:
+    description: use insecure http connection to minio
+    required: false
+  accessKey:
+    description: s3 / minio accessKey
+    required: true
+  secretKey:
+    description: s3 / minio secretKey
+    required: true
+  sessionToken:
+    description: s3 session token - required when using role
+    required: false
+  bucket:
+    description: s3 / minio bucket
+    required: true
+  path:
+    description: "A list of files, directories, and wildcard patterns to cache and restore"
+    required: true
+  key:
+    description: "An explicit key for restoring and saving the cache"
+    required: true
+  restore-keys:
+    description: "An ordered list of keys to use for restoring the cache if no cache hit occurred for key"
+    required: false
+  use-fallback:
+    description: "Use github actions/cache as fallback"
+    required: false
+    default: "true"
+  # zip-option:
+  #   description: zip options
+  #   required: false
+  #   default: -ryq
+  # unzip-option:
+  #   description: unzip options
+  #   required: false
+  #   default: -n
+outputs:
+  cache-hit:
+    description: "A boolean value to indicate an exact match was found for the primary key"
+  cache-size:
+    description: "A integer value denoting the size of the cache object found (measured in bytes)"
+runs:
+  using: node20
+  main: "../dist/restore/index.js"
+branding:
+  icon: "archive"
+  color: "gray-dark"
