+++ src/saveOnly.test.ts
@@ -1,0 +1,5 @@
+import * as minio from "minio";
+
+describe("saveOnly", () => {
+  test("save cache", async () => {});
+});
