+++ src/state.ts
@@ -4,4 +4,5 @@ export enum State {
     "AccessKey" = "access-key",
     "SecretKey" = "secret-key",
     "SessionToken" = "session-token",
+    "Region" = "region"
 }
