+++ src/verify-cache-files.sh
@@ -29,7 +29,7 @@ fi
 
 # Verify file content
 content="$(cat $file)"
-echo "File content:\n$content"
+echo -e "File content:\n$content"
 if [ -z "$(echo $content | grep --fixed-strings "$prefix $GITHUB_RUN_ID")" ]; then
   echo "Unexpected file content"
   exit 1
